improvement: add igniter installer (#208)
with this change, installing live-debugger can be as easy as:

`mix igniter.install live_debugger`

You can generate an example project with live debugger with

`mix igniter.new --with phx.new --install live_debugger`

If this change is accepted, I suggest modifying the readme to have tabs
for setup, one for manual and one for with igniter. You can see how
we do this kind of thing in the Ash getting started guide:

https://hexdocs.pm/ash/get-started.html#create-a-new-project

If this change is accepted we will add a `Live Debugger` option
to the ash-hq.org installer as well.